Move most of the swiper logic into a shared component
